refactor: add separate file for buttons objects
